+++ R/unnest.R
@@ -5,6 +5,7 @@
 #'
 #' @param .df A nested data.table
 #' @param ... Columns to unnest. If empty, unnests all list columns. `tidyselect` compatible.
+#' @param .keep_all Should list columns that were not unnested be kept
 #'
 #' @export
 #' @md
@@ -23,12 +24,12 @@
 #'
 #' nested_df %>%
 #'   unnest.(data, pulled_vec)
-unnest. <- function(.df, ...) {
+unnest. <- function(.df, ..., .keep_all = FALSE) {
   UseMethod("unnest.")
 }
 
 #' @export
-unnest..data.frame <- function(.df, ...) {
+unnest..data.frame <- function(.df, ..., .keep_all = FALSE) {
 
   .df <- as_tidytable(.df)
 
@@ -41,15 +42,20 @@ unnest..data.frame <- function(.df, ...) {
   if (length(dots) == 0) dots <- syms(data_names[list_flag])
   else dots <- select_dots_sym(.df, ...)
 
+  keep_cols <- data_names[!list_flag]
+
+  if (.keep_all) {
+    list_cols <- data_names[list_flag]
+
+    keep_cols <- c(keep_cols, list_cols[list_cols %notin% as.character(dots)])
+  }
+
   unnest_data <- map.(dots, ~ unnest_col(.df, .x))
 
   unnest_nrow <- map_dbl.(unnest_data, nrow)
 
   if (!length(vec_unique(unnest_nrow)) == 1)
     abort("unnested data contains different row counts")
-
-  # Get cols to keep (all non-list cols)
-  keep_cols <- data_names[!list_flag]
 
   # Get number of repeats for keep cols
   rep_vec <- map_dbl.(pull.(.df, !!dots[[1]]), vec_size)
@@ -69,19 +75,21 @@ unnest_col <- function(.df, col = NULL) {
 
   # Check if nested data is a data.frame, data.table, or vector
   nested_data <- pull.(.df, !!col)[[1]]
+  is_vec <- is_bare_vector(nested_data) && !is.matrix(nested_data)
   is_datatable <- is.data.table(nested_data)
-  is_dataframe <- is.data.frame(nested_data)
 
-  if (is_dataframe) {
+  if (is_vec) {
 
+    result_df <- summarize.(.df, !!col := unlist(!!col, recursive = FALSE))
+
+  } else {
+
+    # Convert nested data.frame/tibble/matrix inputs to data.tables
     if (!is_datatable) .df <- mutate.(.df, !!col := map.(!!col, as_tidytable))
 
     result_df <- bind_rows.(pull.(.df, !!col))
-
-  } else {
-    # Unnests a vector
-    result_df <- summarize.(.df, !!col := unlist(!!col, recursive = FALSE))
   }
+
   result_df
 }