+++ tests/testthat/test-unnest.R
@@ -143,3 +143,18 @@ test_that("unnesting works with nested data.table with quosure function", {
   expect_named(unnest_df, c("c","d","a","b"))
   expect_equal(unnest_df$a, start_df$a)
 })
+
+test_that("unnesting works with nested data.table with quosure function", {
+  data_size <- 3
+  list_df <- data.table(test = 1:data_size)
+  test_df <- data.table(x = 1:data_size,
+                        y = replicate(data_size, list_df, simplify = FALSE)) %>%
+    mutate.(z = y)
+
+  result_df <- test_df %>%
+    unnest.(y, .keep_all = TRUE)
+
+  expect_named(result_df, c("x","z","test"))
+  expect_true(is.list(result_df$z))
+  expect_equal(result_df$test, rep(1:3, 3))
+})